mupx v2.1.0 --------------- Add support for DIANA!! Yaaaay!
All new properties:
noMail -> (default: false) if true will not link to mail docker
containersConfig -> (default: false) if true add the support for container server. Container and NGINX config will automatically created with SSL support
containerName -> (default: '') it's mandatory if "containersConfig" is true. It's the name of the container
publishNetwork -> (default: '127.0.0.1' or '0.0.0.0' if "containersConfig" is true)